Revert "Host and Docker Metrics (#2046)"
This reverts commit ecdbc394aed6985abd4adaf529c06a61faf2118c.

+++ edge-util/src/Microsoft.Azure.Devices.Edge.Util/metrics/IMetricsHistogram.cs
@@ -3,6 +3,6 @@ namespace Microsoft.Azure.Devices.Edge.Util.Metrics
 {
     public interface IMetricsHistogram
     {
-        void Update(double value, string[] labelValues);
+        void Update(long value, string[] labelValues);
     }
 }
